third_party/blink/web_tests/wpt_internal/fedcm/fedcm-3pc-blocked-with-iss-3p-ot.sub.https.html

<!DOCTYPE html>
<title>Federated Credential Management API: Test that the API works if third-party cookies are disabled but the ISS origin trial is enabled.</title>
<!--
This is in wpt_internal because this test uses an internal API and also
this is Chrome-specific behavior.
-->
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="https://{{hosts[alt][]}}:{{ports[https][0]}}/wpt_internal/fedcm/resources/third-party-token-injection.js"></script>

<body>
<script type="module">
import {alt_manifest_origin,
        alt_request_options_with_mediation_required,
        fedcm_test,
        fedcm_get_and_select_first_account,
        mark_signed_in} from '/fedcm/support/fedcm-helper.sub.js';

testRunner.setBlockThirdPartyCookies(true);

fedcm_test(async t => {
  await mark_signed_in(alt_manifest_origin);

  const cred = await fedcm_get_and_select_first_account(t, alt_request_options_with_mediation_required());
  assert_equals(cred.token, "token");
}, "Test that we can obtain a token");
</script>
